Initial working commit
This adds the first pass at a UCP module for Puppet. The module does not
attempt to take over management of the individual containers or images,
or to replicate the functionality of the UCP installer. It just provides
a nice Puppet way of downloading and running that installer as part of a
bootstrap process.

Future versions of the module could expand on that, for instance
(optionally) using the Docker module types to ensure the UCP images are
present and that the containers are running. That depends on uptake and
rate-of-change of UCP.